refresh comment number

+++ pages/home/home.js
@@ -26,26 +26,41 @@ Page({
     }
   },
   onLoad: function () {
-    //
-    api.getTopicList().then( resp =>{
-      if (resp.statusCode == 200) {
-        this.setData({
-          posts: resp.data
-        })
-      }
+    // 进入第一次加载
+    api.getTopicList().then( resp => {
+      this.setData({
+        posts: resp.data
+      })
     }).catch(err => {
       console.log("topic", err)
     })
   },
   // 从其它页面返回数据
   onResult: function(data) {
-    if (data && data.ok && data.req == 'newpost') {
-      // data.post
-      // 新增帖子到列表头部
-      this.data.posts.unshift(data.data)
-      this.setData({
-        posts: this.data.posts
-      })
+    if (data && data.ok) {
+      if (data.req == 'newpost') {
+        // data.post
+        // 新增帖子到列表头部
+        this.data.posts.unshift(data.data)
+        this.setData({
+          posts: this.data.posts
+        })
+      } else if (data.req == 'newcomment') {
+        // 刷新评论数量
+        var item = this.data.posts[data.idx]
+        console.log("data:", data)
+        console.log("item:", item)
+        var key = 'posts[' + data.idx + '].stats'
+        if (!item.stats) {
+          item.stats = {}
+        }
+        if (item.stats.comments) {
+          item.stats.comments += 1
+        } else {
+          item.stats.comments = 1
+        }
+        this.setData({ [key]: item.stats })
+      }
     }
     console.log('home, on result data:' + data)
   },
@@ -55,15 +70,10 @@ Page({
   onPullDownRefresh: function () {
       api.getTopicList().then((resp) => {
         wx.stopPullDownRefresh()
-        if (resp.statusCode == 200) {
-          // 
-          this.setData({
-            posts: resp.data
-          })
-
-          console.log(resp.data[0])
-        }
-
+        this.setData({
+          posts: resp.data
+        })
+        console.log(resp.data[0])
       }).catch((err) => {
         wx.stopPullDownRefresh()
         console.log(err)
@@ -85,7 +95,7 @@ Page({
     }
     api.getTopicList(sinceId, limit).then((resp) => {
       this.data.loader.ing = false
-      if (resp.statusCode == 200 && resp.data) {
+      if (resp.data) {
         if(resp.data.length < 20) {
           console.log("no more data..." + sinceId)
           this.data.loader.more = false
@@ -113,9 +123,11 @@ Page({
   },
   commentClick: function(e) {
     var idx = e.currentTarget.dataset.idx
-    console.log("comment clicked", idx)
-    var item = this.data.posts[idx]
-    util.sendRequest('post', item)
+    var post = this.data.posts[idx]
+    util.sendRequest('post', {
+      idx: idx,
+      post: post
+    })
     wx.navigateTo({
       url: '/pages/thread/thread',
     })
@@ -125,24 +137,24 @@ Page({
     var item = this.data.posts[idx]
     var key = 'posts[' + idx + '].stats'
 
-    if (item.stats.favored > 0) {
+    if (!item.stats) {
+      item.stats = { favored: false, favors: 0, comments: 0 }
+    }
+
+    if (item.stats.favored && item.stats.favored > 0) {
       console.log("delete favor")
       api.deletePostFavor(item.id).then( resp => {
-        if (resp.statusCode == 200) {
-          item.stats.favored = false,
-          item.stats.favors -= 1
-          this.setData({[key]: item.stats})
-        }
+        item.stats.favored = false,
+        item.stats.favors -= 1
+        this.setData({[key]: item.stats})
         console.log("delete favor:", resp.statusCode)
       })
     } else {
       console.log("create favor")
       api.createPostFavor(item.id).then((resp) => {
-        if (resp.statusCode == 200) {
-          item.stats.favors += 1
-          item.stats.favored = true
-          this.setData({[key]: item.stats })
-        }
+        item.stats.favors += 1
+        item.stats.favored = true
+        this.setData({[key]: item.stats })
         console.log("favor fail:", resp.statusCode)
       }).catch(err => {
         console.log("favor err:", err)
